udp_server/requirements.txt
asn1crypto==0.24.0
certifi==2018.1.18
chardet==3.0.4
cryptography==2.1.4
idna==2.6
keyring==10.6.0
keyrings.alt==3.0
mysql==0.0.2
mysql-connector==2.2.9
mysqlclient==1.4.6
pyasn1==0.4.8
pycrypto==2.6.1
pygobject==3.26.1
pyxdg==0.25
SecretStorage==2.3.1
six==1.11.0
